git-svn-id: svn+ssh://scm.gforge.inria.fr/svnroot/gtg/branches/gtg_0.2@129 0c38aafd-2db3-40e8-87a3-ec7d7f65c264

+++ inc/OTF_Structs.h
@@ -108,7 +108,7 @@ typedef struct VariableType {
 typedef struct Variable {
     int             parent;
     int             type;
-    int             value;
+    uint64_t        value;
     int             id;
     struct gtg_list token;	/* stored in the variables list */
 } Variable_t;
